+++ src/main/java/com/kaduvill/capnschiselsearch/mixin/client/GuiHitechChiselMixin.java
@@ -1,13 +1,12 @@
 package com.kaduvill.capnschiselsearch.mixin.client;
 
-import com.kaduvill.capnschiselsearch.CapnsChiselSearchConfig;
 import net.minecraft.item.ItemStack;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
 import team.chisel.client.gui.GuiHitechChisel;
 
-@Mixin(value = GuiHitechChisel.class, remap = false)
+@Mixin(value = GuiHitechChisel.class, remap = true)
 public abstract class GuiHitechChiselMixin {
     @Redirect(
             method = "actionPerformed",
@@ -17,10 +16,6 @@ public abstract class GuiHitechChiselMixin {
             )
     )
     private boolean capnschiselsearch$compareHitechTargetIncludingNbt(ItemStack targetStack, ItemStack selectedStack) {
-        if (!CapnsChiselSearchConfig.enableIChiselNbtVariantFix) {
-            return ItemStack.areItemsEqual(targetStack, selectedStack);
-        }
-
         return ItemStack.areItemsEqual(targetStack, selectedStack)
                 && ItemStack.areItemStackTagsEqual(targetStack, selectedStack);
     }