+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.5.0
+
+- Fixed: Chisel recipes breaking after clicking HEI bookmarks
+  - Client-side only
+- Config correctly blocks optional mixins from loading
+
 0.4.0
 - iChisel NBT packet fix
   - Client-side only